Absorb TopNavigation into TaskList
