+++ src/components/App/App.less
@@ -1,6 +1,5 @@
 @import "../../lib/bulma.css";
 @import "https://cdnjs.cloudflare.com/ajax/libs/octicons/3.5.0/octicons.min.css";
-@import "../TopNavigation/TopNavigation.less";
 @import "../NewFilter/NewFilter.less";
 @import "../EditFilter/EditFilter.less";
 @import "../FilterList/FilterList.less";
