--- src/components/TopNavigation/TopNavigation.less
@@ -1,12 +1,0 @@
-.top-navigation {
-  border-bottom: 1px solid rgba(211, 214, 219, 0.5);
-  margin-bottom: 15px;
-
-  .button.is-link {
-    text-decoration: none;
-  }
-
-  .select select {
-    max-width: 295px;
-  }
-}
