--- src/components/TopNavigation/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "TopNavigation",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./TopNavigation.jsx"
-}
